Fix and test INCLUDE_FULL_LIBRARY
The main fix here is to set `REVERSE_DEPS` to `all` when building
with `INCLUDE_FULL_LIBRARY`.

The second fix is to move the definition of `gluOrtho2D` into
`library_glemu.js` which is the only place the `glOrtho` is
defined.  Without this `glOrtho` will show up as an undefined symbol
(at least unless `LEGACY_GL_EMULATION` is set).